Alt Name Reference Dictionary is Working as Intended
Since there are discrepancies between the names of regions in sri_lanka.json and lk.json, the only way to reconcile these discrepancies was to create a dictionary correlating each region name in sri_lanka.json to its corresponding label in lk.json.

After minimal testing, this appears to be a working solution.